+++ Utility/Serpent.h
@@ -30,8 +30,8 @@
 // ** 128bit block size
 // ** 256bit key
 
-extern	void Serpent_set_key(DWORD *l_key,const DWORD *in_key, const DWORD key_len);
-extern	void Serpent_encrypt(const DWORD *l_key,const DWORD *in_blk, DWORD *out_blk);
-extern	void Serpent_decrypt(const DWORD *l_key,const DWORD *in_blk, DWORD *out_blk);
+extern	void Serpent_set_key(unsigned long *l_key,const unsigned long *in_key, const unsigned long key_len);
+extern	void Serpent_encrypt(const unsigned long *l_key,const unsigned long *in_blk, unsigned long *out_blk);
+extern	void Serpent_decrypt(const unsigned long *l_key,const unsigned long *in_blk, unsigned long *out_blk);
 
 #endif
